hashtree (plum_db v0.3.10)

This module implements a persistent, on-disk hash tree that is used predominately for active anti-entropy exchange in Riak. The tree consists of two parts, a set of unbounded on-disk segments and a fixed size hash tree (that may be on-disk or in-memory) constructed over these segments.

A graphical description of this design can be found in: docs/hashtree.md

Each segment logically represents an on-disk list of (key, hash) pairs. Whereas the hash tree is represented as a set of levels and buckets, with a fixed width (or fan-out) between levels that determines how many buckets of a child level are grouped together and hashed to represent a bucket at the parent level. Each leaf in the tree corresponds to a hash of one of the on-disk segments. For example, a tree with a width of 4 and 16 segments would look like the following:

  %%{init: {'theme': 'neutral' } }%%
  graph TD
    r[" "]
 
    a[" "]; b[" "]; c[" "]; d[" "]
 
    a1[" "]; a2[" "]; a3[" "]; a4[" "];
    a11[" "]; a12[" "]; a13[" "]; a14[" "]; a15[" "]; a16[" "];
    a21[" "]; a22[" "]; a23[" "]; a24[" "]; a25[" "]; a26[" "];
    a31[" "]; a32[" "]; a33[" "]; a34[" "]; a35[" "]; a36[" "];
    a41[" "]; a42[" "]; a43[" "]; a44[" "]; a45[" "]; a46[" "];
 
    b1[" "]; b2[" "]; b3[" "]; b4[" "];
    b11[" "]; b12[" "]; b13[" "]; b14[" "]; b15[" "]; b16[" "];
    b21[" "]; b22[" "]; b23[" "]; b24[" "]; b25[" "]; b26[" "];
    b31[" "]; b32[" "]; b33[" "]; b34[" "]; b35[" "]; b36[" "];
    b41[" "]; b42[" "]; b43[" "]; b44[" "]; b45[" "]; b46[" "];
 
    c1[" "]; c2[" "]; c3[" "]; c4[" "];
    c11[" "]; c12[" "]; c13[" "]; c14[" "]; c15[" "]; c16[" "];
    c21[" "]; c22[" "]; c23[" "]; c24[" "]; c25[" "]; c26[" "];
    c31[" "]; c32[" "]; c33[" "]; c34[" "]; c35[" "]; c36[" "];
    c41[" "]; c42[" "]; c43[" "]; c44[" "]; c45[" "]; c46[" "];
 
    d1[" "]; d2[" "]; d3[" "]; d4[" "];
    d11[" "]; d12[" "]; d13[" "]; d14[" "]; d15[" "]; d16[" "];
    d21[" "]; d22[" "]; d23[" "]; d24[" "]; d25[" "]; d26[" "];
    d31[" "]; d32[" "]; d33[" "]; d34[" "]; d35[" "]; d36[" "];
    d41[" "]; d42[" "]; d43[" "]; d44[" "]; d45[" "]; d46[" "];
 
    subgraph " "
    a; b; c; d
    end
 
    r --> a; r --> b; r --> c; r --> d
 
    subgraph A [" "]
    a2; a2; a3; a4
    end
 
    subgraph A1 [" "]
    a11; a12; a13; a14; a15; a16
    end
    subgraph A2 [" "]
    a21; a22; a23; a24; a25; a26
    end
    subgraph A3 [" "]
    a31; a32; a33; a34; a35; a36
    end
    subgraph A4 [" "]
    a41; a42; a43; a44; a45; a46
    end
 
    a --> a1; a --> a2; a --> a3; a --> a4
    a1 --- a11 --- a12 --- a13 --- a14 --- a15 --- a16
    a2 --- a21 --- a22 --- a23 --- a24 --- a25 --- a26
    a3 --- a31 --- a32 --- a33 --- a34 --- a35 --- a36
    a4 --- a41 --- a42 --- a43 --- a44 --- a45 --- a46
 
    subgraph B [" "]
    b1; b2; b3; b4
    end
 
    subgraph B1 [" "]
    b11; b12; b13; b14; b15; b16
    end
    subgraph B2 [" "]
    b21; b22; b23; b24; b25; b26
    end
    subgraph B3 [" "]
    b31; b32; b33; b34; b35; b36
    end
    subgraph B4 [" "]
    b41; b42; b43; b44; b45; b46
    end
 
    b --> b1; b --> b2; b --> b3; b --> b4
    b1 --- b11 --- b12 --- b13 --- b14 --- b15 --- b16
    b2 --- b21 --- b22 --- b23 --- b24 --- b25 --- b26
    b3 --- b31 --- b32 --- b33 --- b34 --- b35 --- b36
    b4 --- b41 --- b42 --- b43 --- b44 --- b45 --- b46
 
    subgraph C [" "]
    c1; c2; c3; c4
    end
 
    subgraph C1 [" "]
    c11; c12; c13; c14; c15; c16
    end
    subgraph C2 [" "]
    c21; c22; c23; c24; c25; c26
    end
    subgraph C3 [" "]
    c31; c32; c33; c34; c35; c36
    end
    subgraph C4 [" "]
    c41; c42; c43; c44; c45; c46
    end
 
    c --> c1; c --> c2; c --> c3; c --> c4
    c1 --- c11 --- c12 --- c13 --- c14 --- c15 --- c16
    c2 --- c21 --- c22 --- c23 --- c24 --- c25 --- c26
    c3 --- c31 --- c32 --- c33 --- c34 --- c35 --- c36
    c4 --- c41 --- c42 --- c43 --- c44 --- c45 --- c46
 
    subgraph D [" "]
    d1; d2; d3; d4
    end
 
    subgraph D1 [" "]
    d11; d12; d13; d14; d15; d16
    end
    subgraph D2 [" "]
    d21; d22; d23; d24; d25; d26
    end
    subgraph D3 [" "]
    d31; d32; d33; d34; d35; d36
    end
    subgraph D4 [" "]
    d41; d42; d43; d44; d45; d46
    end
 
    d --> d1; d --> d2; d --> d3; d --> d4
    d1 --- d11 --- d12 --- d13 --- d14 --- d15 --- d16
    d2 --- d21 --- d22 --- d23 --- d24 --- d25 --- d26
    d3 --- d31 --- d32 --- d33 --- d34 --- d35 --- d36
    d4 --- d41 --- d42 --- d43 --- d44 --- d45 --- d46
 
  

level buckets 1: [0] 2: [0 1 2 3] 3: [0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15]

With each bucket entry of the form {bucket-id, hash}, eg. {0, binary()}. The hash for each of the entries at level 3 would come from one of the 16 segments, while the hashes for entries at level 1 and 2 are derived from the lower levels.

Specifically, the bucket entries in level 2 would come from level 3: 0: hash([ 0 1 2 3]) 1: hash([ 4 5 6 7]) 2: hash([ 8 9 10 11]) 3: hash([12 13 14 15])

And the bucket entries in level 1 would come from level 2: 1: hash([hash([ 0 1 2 3]) hash([ 4 5 6 7]) hash([ 8 9 10 11]) hash([12 13 14 15])])

When a (key, hash) pair is added to the tree, the key is hashed to determine which segment it belongs to and inserted/upserted into the segment. Rather than update the hash tree on every insert, a dirty bit is set to note that a given segment has changed. The hashes are then updated in bulk before performing a tree exchange

To update the hash tree, the code iterates over each dirty segment, building a list of (key, hash) pairs. A hash is computed over this list, and the leaf node in the hash tree corresponding to the given segment is updated. After iterating over all dirty segments, and thus updating all leaf nodes, the update then continues to update the tree bottom-up, updating only paths that have changed. As designed, the update requires a single sparse scan over the on-disk segments and a minimal traversal up the hash tree.

The heavy-lifting of this module is provided by LevelDB. What is logically viewed as sorted on-disk segments is in reality a range of on-disk (segment, key, hash) values written to LevelDB. Each insert of a (key, hash) pair therefore corresponds to a single LevelDB write (no read necessary). Likewise, the update operation is performed using LevelDB iterators.

When used for active anti-entropy in Riak, the hash tree is built once and then updated in real-time as writes occur. A key design goal is to ensure that adding (key, hash) pairs to the tree is non-blocking, even during a tree update or a tree exchange. This is accomplished using LevelDB snapshots. Inserts into the tree always write directly to the active LevelDB instance, however updates and exchanges operate over a snapshot of the tree.

In order to improve performance, writes are buffered in memory and sent to LevelDB using a single batch write. Writes are flushed whenever the buffer becomes full, as well as before updating the hashtree.
Tree exchange is provided by the compare/4 function. The behavior of this function is determined through a provided function that implements logic to get buckets and segments for a given remote tree, as well as a callback invoked as key differences are determined. This generic interface allows for tree exchange to be implemented in a variety of ways, including directly against to local hash tree instances, over distributed Erlang, or over a custom protocol over a TCP socket. See local_compare/2 and do_remote/1 for examples (-ifdef(TEST) only).

Types

acc_fun/1

Specs

acc_fun(Acc) :: fun(([keydiff()], Acc) -> Acc).

bucket_bin/0

Specs

bucket_bin() :: <<_:320>>.

hashtree/0
(opaque)

Specs

hashtree()

index/0

Specs

index() :: non_neg_integer().

keydiff/0

Specs

keydiff() :: {missing | remote_missing | different, binary()}.

meta_bin/0

Specs

meta_bin() :: <<_:8, _:_*8>>.

orddict/0

Specs

orddict() :: orddict:orddict().

proplist/0

Specs

proplist() :: proplists:proplist().

remote_fun/0

Specs

remote_fun() ::
    fun((get_bucket | key_hashes | init | final, {integer(), integer()} | integer() | term()) ->
            any()).

segment_bin/0

Specs

segment_bin() :: <<_:256, _:_*8>>.

select_fun/1

Specs

select_fun(T) :: fun((orddict()) -> T).

tree_id_bin/0

Specs

tree_id_bin() :: <<_:176>>.

Functions

close(State)

Specs

close(hashtree()) -> hashtree().

compare(Tree, Remote, AccFun, Acc)

delete(Key, State)

Specs

delete(binary(), hashtree()) -> hashtree().

destroy(Path)

Specs

destroy(string() | hashtree()) -> ok | hashtree().

flush_buffer(State)

get_bucket(Level, Bucket, State)

Specs

get_bucket(integer(), integer(), hashtree()) -> orddict().

insert(Key, ObjHash, State)

Specs

insert(binary(), binary(), hashtree()) -> hashtree().

insert(Key, ObjHash, State, Opts)

Specs

insert(binary(), binary(), hashtree(), proplist()) -> hashtree().

key_hashes(State, Segment)

Specs

key_hashes(hashtree(), integer()) -> [{integer(), orddict()}].

levels(State)

Specs

levels(hashtree()) -> pos_integer().

mem_levels(State)

Specs

mem_levels(hashtree()) -> integer().

new()

Specs

new() -> hashtree() | no_return().

new(TreeId, Options)

Specs

new({index(), tree_id_bin() | non_neg_integer()}, proplist()) -> hashtree() | no_return();
   ({index(), tree_id_bin() | non_neg_integer()}, hashtree()) -> hashtree() | no_return().

new(_, LinkedStore, Options)

Specs

new({index(), tree_id_bin() | non_neg_integer()}, hashtree(), proplist()) ->
       hashtree() | no_return().

read_meta(Key, State)

Specs

read_meta(binary(), hashtree()) -> {ok, binary()} | undefined.

rehash_tree(State)

Specs

rehash_tree(hashtree()) -> hashtree().

segments(State)

Specs

segments(hashtree()) -> pos_integer().

top_hash(State)

Specs

top_hash(hashtree()) -> [] | [{0, binary()}].

update_perform(State2)

Specs

update_perform(hashtree()) -> hashtree().

update_snapshot(State)

Specs

update_snapshot(hashtree()) -> {hashtree(), hashtree()}.

update_tree(State)

Specs

update_tree(hashtree()) -> hashtree().

width(State)

Specs

width(hashtree()) -> pos_integer().

write_meta(Key, Value, State)

Specs

write_meta(binary(), binary(), hashtree()) -> hashtree().