plum_db (plum_db v0.3.10)

Types

continuation/0
(opaque)

Specs

continuation()

continuation_or_eot/0

Specs

continuation_or_eot() :: continuation() | eot().

delete_opts/0

Specs

delete_opts() :: [].

eot/0

Specs

eot() :: '$end_of_table'.

erase_opts/0

Specs

erase_opts() :: [].

fold_elements_fun/0

Specs

fold_elements_fun() :: fun(({plum_db_key(), plum_db_object()}, any()) -> any()).

fold_fun/0

Specs

fold_fun() :: fun(({plum_db_key(), value_or_values()}, any()) -> any()).

fold_opts/0

Specs

fold_opts() :: it_opts().

foreach_fun/0

Specs

foreach_fun() :: fun(({plum_db_key(), value_or_values()}) -> any()).

get_opt/0

Specs

get_opt() :: get_opt_default_val() | get_opt_resolver() | get_opt_allow_put().

get_opt_allow_put/0

Specs

get_opt_allow_put() :: {allow_put, boolean()}.

get_opt_default_val/0

Specs

get_opt_default_val() :: {default, plum_db_value()}.

get_opt_resolver/0

Specs

get_opt_resolver() :: {resolver, plum_db_resolver()}.

get_opts/0

Specs

get_opts() :: [get_opt()].

it_opt/0

Specs

it_opt() ::
    it_opt_resolver() |
    it_opt_first() |
    it_opt_default() |
    it_opt_keymatch() |
    it_opt_keys_only() |
    it_opt_partitions().

it_opt_default/0

Specs

it_opt_default() :: {default, plum_db_value() | it_opt_default_fun()}.

it_opt_default_fun/0

Specs

it_opt_default_fun() :: fun((plum_db_key()) -> plum_db_value()).

it_opt_first/0

Specs

it_opt_first() :: {first, term()}.

it_opt_keymatch/0

Specs

it_opt_keymatch() :: {match, term()}.

it_opt_keys_only/0

Specs

it_opt_keys_only() :: {keys_only, boolean()}.

it_opt_partitions/0

Specs

it_opt_partitions() :: {partitions, [partition()]}.

it_opt_resolver/0

Specs

it_opt_resolver() :: {resolver, plum_db_resolver() | lww}.

it_opts/0

Specs

it_opts() :: [it_opt()].

iterator/0
(opaque)

Specs

iterator()

iterator_element/0

Specs

iterator_element() :: {plum_db_pkey(), plum_db_object()}.

match_opt_limit/0

Specs

match_opt_limit() :: pos_integer() | infinity.

match_opt_remove_tombstones/0

Specs

match_opt_remove_tombstones() :: boolean().

match_opts/0

Specs

match_opts() :: [it_opt() | match_opt_limit() | match_opt_remove_tombstones()].

partition/0

Specs

partition() :: non_neg_integer().

plum_db_broadcast/0

Specs

plum_db_broadcast() :: #plum_db_broadcast{}.

plum_db_context/0

Specs

plum_db_context() :: plum_db_dvvset:vector().

plum_db_key/0

Specs

plum_db_key() :: any().

plum_db_modifier/0

Specs

plum_db_modifier() ::
    fun(([plum_db_value() | plum_db_tombstone()] | undefined) -> plum_db_value()).

plum_db_object/0

Specs

plum_db_object() :: {object, plum_db_dvvset:clock()}.

plum_db_pkey/0

Specs

plum_db_pkey() :: {plum_db_prefix(), plum_db_key()}.

plum_db_pkey_pattern/0

Specs

plum_db_pkey_pattern() :: {plum_db_prefix_pattern(), plum_db_key() | plum_db_wildcard()}.

plum_db_prefix/0

Specs

plum_db_prefix() :: {binary() | atom(), binary() | atom()}.

plum_db_prefix_pattern/0

Specs

plum_db_prefix_pattern() ::
    {binary() | atom() | plum_db_wildcard(), binary() | atom() | plum_db_wildcard()}.

plum_db_resolver/0

Specs

plum_db_resolver() ::
    fun((plum_db_key() | plum_db_pkey(), plum_db_value() | plum_db_tombstone()) -> plum_db_value()).

plum_db_tombstone/0

Specs

plum_db_tombstone() :: '$deleted'.

plum_db_value/0

Specs

plum_db_value() :: any().

plum_db_wildcard/0

Specs

plum_db_wildcard() :: '_'.

prefix_type/0

Specs

prefix_type() :: ram | ram_disk | disk.

prefixes/0

Specs

prefixes() :: #{binary() | atom() => prefix_type()}.

put_opts/0

Specs

put_opts() :: [].

remote_iterator/0

Specs

remote_iterator() :: #remote_iterator{}.

state/0

Specs

state() :: #state{}.

value_or_values/0

Specs

value_or_values() ::
    [plum_db_value() | plum_db_tombstone()] | plum_db_value() | plum_db_tombstone().

Functions

broadcast_data(Plum_db_broadcast)

Specs

broadcast_data(plum_db_broadcast()) -> {{plum_db_pkey(), plum_db_context()}, plum_db_object()}.

Deconstructs a broadcast that is sent using broadcast/2 as the handling module returning the message id and payload.
> This function is part of the implementation of the partisan_plumtree_broadcast_handler behaviour. > You should never call it directly.

delete(FullPrefix, Key)

Specs

delete(plum_db_prefix(), plum_db_key()) -> ok.

Same as delete(FullPrefix, Key, [])

delete(FullPrefix, Key, Opts)

Specs

delete(plum_db_prefix(), plum_db_key(), delete_opts()) -> ok.

Logically deletes the value associated with the given prefix and key locally and then triggers a broradcast to notify other nodes in the cluster. Currently there are no delete options.
NOTE: currently deletion is logical and no GC is performed.

erase(FullPrefix, Key)

Specs

erase(plum_db_prefix(), plum_db_key()) -> ok.

Same as delete(FullPrefix, Key, [])

erase(PKey, Key, Opts)

Specs

erase(plum_db_prefix(), plum_db_key(), erase_opts()) -> ok.

Logically erases the value associated with the given prefix and key locally and then triggers a broradcast to notify other nodes in the cluster. Currently there are no erase options.
NOTE: currently deletion is logical and no GC is performed.

exchange(Peer)

Specs

exchange(node()) -> {ok, pid()} | {error, term()}.

Triggers an asynchronous exchange. Calls exchange/2 with an empty map as the second argument. > The exchange is only triggered if the application option aae_enabled is set to true.

exchange(Peer, Opts)

Specs

exchange(node(), map()) -> {ok, pid()} | {error, term()}.

Triggers an asynchronous exchange. The exchange is performed asynchronously by spawning a supervised process. Read the plum_db_exchanges_sup documentation.

Opts is a map accepting the following options:

* timeout (milliseconds) –– timeout for the AAE exchange to conclude.
> The exchange is only triggered if the application option aae_enabled is set to true.

fold(Fun, Acc0, PrefixPatternOrCont)

Specs

fold(Fun :: fold_fun(),
     Acc0 :: any(),
     PrefixPatternOrCont :: plum_db_prefix_pattern() | continuation_or_eot()) ->
        any() | {any(), continuation_or_eot()}.

Same as fold(Fun, Acc0, FullPrefix, []).

fold(Fun, Acc0, PrefixPatternOrCont, Opts)

Specs

fold(Fun :: fold_fun(),
     Acc0 :: any(),
     PrefixPatternOrCont :: plum_db_prefix_pattern() | continuation_or_eot(),
     Opts :: fold_opts()) ->
        any() | {any(), continuation_or_eot()}.

Fold over all keys and values stored under a given prefix/subprefix. Available options are the same as those provided to iterator/2. To return early, throw {break, Result} in your fold function.

fold_elements(Fun, Acc0, FullPrefix)

Specs

fold_elements(fold_elements_fun(), any(), plum_db_prefix()) -> any().

Same as fold_elements(Fun, Acc0, FullPrefix, []).

fold_elements(Fun, Acc0, FullPrefix, Opts)

Specs

fold_elements(fold_elements_fun(), any(), plum_db_prefix(), fold_opts()) -> any().

Fold over all elements stored under a given prefix/subprefix. Available options are the same as those provided to iterator/2. To return early, throw {break, Result} in your fold function.

foreach(Fun, FullPrefixPattern)

Specs

foreach(foreach_fun(), plum_db_prefix_pattern()) -> any().

Same as fold(Fun, Acc0, FullPrefix, []).

foreach(Fun, FullPrefixPattern, Opts)

Specs

foreach(foreach_fun(), plum_db_prefix_pattern(), fold_opts()) -> any().

Fold over all keys and values stored under a given prefix/subprefix. Available options are the same as those provided to iterator/2. To return early, throw {break, Result} in your fold function.

get(FullPrefix, Key)

Specs

get(plum_db_prefix(), plum_db_key()) -> plum_db_value() | undefined.

Same as get(FullPrefix, Key, [])

get(FullPrefix, Key, Opts)

Specs

get(plum_db_prefix(), plum_db_key(), get_opts()) -> plum_db_value() | undefined.

Retrieves the local value stored at the given fullprefix FullPrefix and key Key using options Opts.

Returns the stored value if found. If no value is found and Opts contains a value for the key default, this value is returned. Otherwise returns the atom undefined.

Opts is a property list that can take the following options:

* default – value to return if no value is found, Defaults to undefined. * resolver – The atom lww or a plum_db_resolver() that resolves conflicts if they are encountered. Defaults to lww (last-write-wins). * allow_put – whether or not to write and broadcast a resolved value. Defaults to true.

Example: Simple get

  > plum_db:get({foo, a}, x).
  undefined.
  > plum_db:get({foo, a}, x, [{default, 1}]).
  1.
  > plum_db:put({foo, a}, x, 100).
  ok
  > plum_db:get({foo, a}, x).
  100.

Example: Resolving with a custom function

  Fun = fun(A, B) when A > B -> A; _ -> B end,
  > plum_db:get({foo, a}, x, [{resolver, Fun}]).
> NOTE: an update will be broadcasted if conflicts are resolved and allow_put is true. However, any further conflicts generated by concurrent writes during resolution are not resolved.

get_object(PKey)

Returns a Dotted Version Vector Set or undefined. When reading the value for a subsequent call to put/3 the context can be obtained using plum_db_object:context/1. Values can obtained w/ plum_db_object:values/1.

get_object(PKey, Opts)

Returns a Dotted Version Vector Set or undefined. When reading the value for a subsequent call to put/3 the context can be obtained using plum_db_object:context/1. Values can obtained w/ plum_db_object:values/1.

get_object(Node, PKey, Opts)

Specs

get_object(node(), plum_db_pkey(), get_opts()) -> plum_db_object() | undefined.

Same as get/1 but reads the value from Node This is function is used by plum_db_exchange_statem.

graft(_)

Specs

graft({plum_db_pkey(), plum_db_context()}) -> stale | {ok, plum_db_object()} | {error, term()}.

Returns the object associated with the given prefixed key Pkey and context Context (message id) if the currently stored version has an equal context. Otherwise returns the atom stale.

Because it assumes that a grafted context can only be causally older than the local view, a stale response means there is another message that subsumes the grafted one.
> This function is part of the implementation of the partisan_plumtree_broadcast_handler behaviour. > You should never call it directly.

is_partition(Id)

Specs

is_partition(partition()) -> boolean().

Returns true if an identifier is a valid partition.

is_stale(_)

Specs

is_stale({plum_db_pkey(), plum_db_context()}) -> boolean().

Returns false if the update (or a causally newer update) has already been received (stored locally).
> This function is part of the implementation of the partisan_plumtree_broadcast_handler behaviour. > You should never call it directly.

iterate(Remote_iterator)

Specs

iterate(iterator() | remote_iterator()) -> iterator() | remote_iterator().

Advances the iterator by one key, full-prefix or sub-prefix

iterator()

Specs

iterator() -> iterator().

Returns a full-prefix iterator: an iterator for all full-prefixes that have keys stored under them. When done with the iterator, iterator_close/1 must be called. This iterator works across all existing store partitions, treating the set of partitions as a single logical database. As a result, ordering is partial per partition and not global across them.
Same as calling iterator({undefined, undefined}).

iterator(Term)

Specs

iterator(plum_db_prefix() | continuation()) -> iterator().

Same as calling iterator(FullPrefix, []).

iterator(Continuation, Opts)

Specs

iterator(plum_db_prefix_pattern() | continuation(), it_opts()) -> iterator().

Return an iterator pointing to the first key stored under a prefix

This function can take the following options:
* resolver: either the atom lww or a function that resolves conflicts if they are encounted (see get/3 for more details). Conflict resolution is performed when values are retrieved (see iterator_key_value/1 and iterator_key_values/1). If no resolver is provided no resolution is performed. The default is to not provide a resolver. * allow_put: whether or not to write and broadcast a resolved value. defaults to true. * default: Used when the value an iterator points to is a tombstone. default is either an arity-1 function or a value. If a function, the key the iterator points to is passed as the argument and the result is returned in place of the tombstone. If default is a value, the value is returned in place of the tombstone. This applies when using functions such as iterator_key_values/1 and iterator_key_values/1. * first - the key this iterator should start at, equivalent to calling iterator_move/2 passing the key as the second argument. * match: If match is undefined then all keys will may be visted by the iterator, match can be: * an erlang term - which will be matched exactly against a key * '_' - the wildcard term which matches anything * an erlang tuple containing terms and '_' - if tuples are used as keys this can be used to iterate over some subset of keys * partitions: The list of partitions this iterator should cover. If undefined it will cover all partitions (pdb:partitions/0) * keys_only: wether to iterate only on keys (default: false)

iterator_close(Remote_iterator)

Specs

iterator_close(iterator() | iterator() | remote_iterator()) -> ok.

Closes the iterator. This function must be called on all open iterators

iterator_default(Iterator)

Specs

iterator_default(iterator() | iterator()) ->
                    plum_db_tombstone() | plum_db_value() | it_opt_default_fun().

Returns the value returned when an iterator points to a tombstone. If the default used when creating the given iterator is a function it will be applied to the current key the iterator points at. If no default was provided the tombstone value was returned. This function should only be called after checking iterator_done/1.

iterator_done(Remote_iterator)

Specs

iterator_done(iterator() | iterator() | remote_iterator()) -> boolean().

Returns true if there is nothing more to iterate over

iterator_element(Remote_iterator)

Specs

iterator_element(iterator() | iterator() | remote_iterator()) -> iterator_element().

iterator_key(Iterator)

Specs

iterator_key(iterator()) -> plum_db_key() | undefined.

Return the key pointed at by the iterator. Before calling this function, check the iterator is not complete w/ iterator_done/1. No conflict resolution will be performed as a result of calling this function.

iterator_key_value(Iterator)

Specs

iterator_key_value(iterator() | remote_iterator() | iterator()) ->
                      {plum_db_key(), plum_db_value()} | {error, conflict}.

Returns a single value pointed at by the iterator. If there are conflicts and a resolver was specified in the options when creating this iterator, they will be resolved. Otherwise, and error is returned. If conflicts are resolved, the resolved value is written locally and a broadcast is performed to update other nodes in the cluster if allow_put is true (the default value). If allow_put is false, values are resolved but not written or broadcast.
NOTE: if resolution may be performed this function must be called at most once before calling iterate/1 on the iterator (at which point the function can be called once more).

iterator_key_values(Iterator)

Specs

iterator_key_values(iterator()) -> {plum_db_key(), value_or_values()}.

Return the key and all sibling values pointed at by the iterator. Before calling this function, check the iterator is not complete w/ iterator_done/1. If a resolver was passed to iterator/0 when creating the given iterator, siblings will be resolved using the given function or last-write-wins (if lww is passed as the resolver). If no resolver was used then no conflict resolution will take place. If conflicts are resolved, the resolved value is written to local store and a broadcast is submitted to update other nodes in the cluster if allow_put is true. If allow_put is false the values are resolved but are not written or broadcast. A single value is returned as the second element of the tuple in the case values are resolved. If no resolution takes place then a list of values will be returned as the second element (even if there is only a single sibling).
NOTE: if resolution may be performed this function must be called at most once before calling iterate/1 on the iterator (at which point the function can be called once more).

iterator_prefix(Remote_iterator)

Specs

iterator_prefix(iterator() | remote_iterator()) -> plum_db_prefix().

Returns the full-prefix being iterated by this iterator.

match(Cont)

Specs

match(continuation_or_eot()) ->
         {[{plum_db_key(), value_or_values()}], continuation_or_eot()} | '$end_of_table'.

match(Continuation, Opts)

Specs

match(continuation_or_eot(), match_opts()) ->
         {[{plum_db_key(), value_or_values()}], continuation()} | '$end_of_table'.

match(FullPrefix0, KeyPattern, Opts0)

Specs

match(plum_db_prefix_pattern(), plum_db_pkey_pattern(), match_opts()) ->
         [{plum_db_key(), value_or_values()}] |
         {[{plum_db_key(), value_or_values()}], continuation()} |
         '$end_of_table'.

merge(_, Obj)

Specs

merge({plum_db_pkey(), undefined | plum_db_context()}, undefined | plum_db_object()) -> boolean().

Merges a remote copy of an object record sent via broadcast w/ the local view for the key contained in the message id. If the remote copy is causally older than the current data stored then false is returned and no updates are merged. Otherwise, the remote copy is merged (possibly generating siblings) and true is returned.
> This function is part of the implementation of the partisan_plumtree_broadcast_handler behaviour. > You should never call it directly.

merge(Node, _, Obj)

Specs

merge(node(), {plum_db_pkey(), undefined | plum_db_context()}, plum_db_object()) -> boolean().

Same as merge/2 but merges the object on Node
> This function is part of the implementation of the partisan_plumtree_broadcast_handler behaviour. > You should never call it directly.

partition_count()

Specs

partition_count() -> non_neg_integer().

Returns the number of partitions.

partitions()

Specs

partitions() -> [partition()].

Returns the list of the partition identifiers starting at 0.

prefix_hash(Partition, Prefix)

Specs

prefix_hash(partition(), plum_db_prefix()) -> binary() | undefined.

Return the local hash associated with a full-prefix or prefix. The hash value is updated periodically and does not always reflect the most recent value. This function can be used to determine when keys stored under a full-prefix or prefix have changed. If the tree has not yet been updated or there are no keys stored the given (full-)prefix. undefined is returned.

prefix_type(Prefix)

Specs

prefix_type(term()) -> prefix_type() | undefined.

prefixes()

Specs

prefixes() -> prefixes().

Returns a mapping of prefixes (the first element of a plum_db_prefix() tuple) to prefix_type() only for those prefixes for which a type was declared using the application optiont prefixes.

put(FullPrefix, Key, ValueOrFun)

Specs

put(plum_db_prefix(), plum_db_key(), plum_db_value() | plum_db_modifier()) -> ok.

Same as put(FullPrefix, Key, Value, [])

put(FullPrefix, Key, ValueOrFun, Opts)

Specs

put(plum_db_prefix(), plum_db_key(), plum_db_value() | plum_db_modifier(), put_opts()) -> ok.

Stores or updates the value at the given prefix and key locally and then triggers a broadcast to notify other nodes in the cluster. Currently, there are no put options.
NOTE: because the third argument to this function can be a plum_db_modifier(), used to resolve conflicts on write, values cannot be functions. To store functions wrap them in another type like a tuple.

remote_iterator(Node)

Specs

remote_iterator(node()) -> remote_iterator().

Create an iterator on Node. This allows for remote iteration by having the worker keep track of the actual iterator (since ets continuations cannot cross node boundaries). The iterator created iterates all full-prefixes. Once created the rest of the iterator API may be used as usual. When done with the iterator, iterator_close/1 must be called

remote_iterator(Node, FullPrefix)

Specs

remote_iterator(node(), plum_db_prefix()) -> remote_iterator().

start_link()

Specs

start_link() -> {ok, pid()} | ignore | {error, term()}.

Start the plum_db server and link to calling process. The plum_db server is responsible for managing local and remote iterators. No API function uses the server itself.

sync_exchange(Peer)

Specs

sync_exchange(node()) -> ok | {error, term()}.

Triggers a synchronous exchange. Calls sync_exchange/2 with an empty map as the second argument. > The exchange is only triggered if the application option aae_enabled is set to true.

sync_exchange(Peer, Opts0)

Specs

sync_exchange(node(), map()) -> ok | {error, term()}.

Triggers a synchronous exchange. The exchange is performed synchronously by spawning a supervised process and waiting (blocking) till it finishes. Read the plum_db_exchanges_sup documentation.

Opts is a map accepting the following options:

* timeout (milliseconds) –– timeout for the AAE exchange to conclude.
> The exchange is only triggered if the application option aae_enabled is set to true.

take(FullPrefix, Key)

Specs

take(plum_db_prefix(), plum_db_key()) -> plum_db_value() | undefined.

take(FullPrefix, Key, Opts)

Specs

take(plum_db_prefix(), plum_db_key(), get_opts()) -> plum_db_value() | undefined.

to_list(FullPrefix)

Specs

to_list(plum_db_prefix()) -> [{plum_db_key(), value_or_values()}].

Same as to_list(FullPrefix, [])

to_list(FullPrefix, Opts)

Specs

to_list(FullPrefix :: plum_db_prefix(), Opts :: fold_opts()) ->
           [{plum_db_key(), value_or_values()}].

Return a list of all keys and values stored under a given prefix/subprefix. Available options are the same as those provided to iterator/2.