plum_db_dvvset (plum_db v0.3.10)

An Erlang implementation of *compact* Dotted Version Vectors, which provides a container for a set of concurrent values (siblings) with causal order information.
For further reading, visit the github page.

Types

clock/0

Specs

clock() :: {entries(), values()}.

counter/0

Specs

counter() :: non_neg_integer().

entries/0

Specs

entries() :: [{id(), counter(), values()}].

id/0

Specs

id() :: any().

value/0

Specs

value() :: any().

values/0

Specs

values() :: [value()].

vector/0

Specs

vector() :: [{id(), counter()}].

Functions

equal(C1, C2)

Specs

equal(clock() | vector(), clock() | vector()) -> boolean().

Compares the equality of both clocks, regarding only the causal histories, thus ignoring the values.

ids(_)

Specs

ids(clock()) -> [id()].

Returns all the ids used in this clock set.

join(_)

Specs

join(clock()) -> vector().

Return a version vector that represents the causal history.

last(LessOrEqual, C)

Specs

last(LessOrEqual :: fun((value(), value()) -> boolean()), clock()) -> value().

Returns the latest value in the clock set, according to function F(A,B), which returns *true* if A compares less than or equal to B, false otherwise.

less(_, _)

Specs

less(clock(), clock()) -> boolean().

Returns True if the first clock is causally older than the second clock, thus values on the first clock are outdated. Returns False otherwise.

lww(LessOrEqual, C)

Specs

lww(LessOrEqual :: fun((value(), value()) -> boolean()), clock()) -> clock().

Return a clock with the same causal history, but with only one value in its original position. This value is the newest value in the given clock, according to function F(A,B), which returns *true* if A compares less than or equal to B, false otherwise.

map(F, _)

Specs

map(fun((value()) -> value()), clock()) -> clock().

Maps (applies) a function on all values in this clock set, returning the same clock set with the updated values.

new(Vs)

Specs

new(value() | [value()]) -> clock().

Constructs a new clock set without causal history, and receives a list of values that gos to the anonymous list.

new(VV, Vs)

Specs

new(vector(), value() | [value()]) -> clock().

Constructs a new clock set with the causal history of the given version vector / vector clock, and receives a list of values that gos to the anonymous list. The version vector SHOULD BE a direct result of join/1.

reconcile(Winner, C)

Specs

reconcile(Winner :: fun(([value()]) -> value()), clock()) -> clock().

Return a clock with the same causal history, but with only one value in the anonymous placeholder. This value is the result of the function F, which takes all values and returns a single new value.

size(_)

Specs

size(clock()) -> non_neg_integer().

Returns the total number of values in this clock set.

sync(L)

Specs

sync([clock()]) -> clock().

Synchronizes a list of clocks using sync/2. It discards (causally) outdated values, while merging all causal histories.

update(_, I)

Specs

update(clock(), id()) -> clock().

Advances the causal history with the given id. The new value is the *anonymous dot* of the clock. The client clock SHOULD BE a direct result of new/2.

update(_, Cr, I)

Specs

update(clock(), clock(), id()) -> clock().

Advances the causal history of the first clock with the given id, while synchronizing with the second clock, thus the new clock is causally newer than both clocks in the argument. The new value is the *anonymous dot* of the clock. The first clock SHOULD BE a direct result of new/2, which is intended to be the client clock with the new value in the *anonymous dot* while the second clock is from the local server.

values(_)

Specs

values(clock()) -> [value()].

Returns all the values used in this clock set, including the anonymous values.