plum_db_events (plum_db v0.3.10)

Functions

add_callback(Fn)

Specs

add_callback(fun((any()) -> any())) -> {ok, reference()}.

Adds a callback function. The function needs to have a single argument representing the event that has been fired.

add_handler(Handler, Args)

Adds an event handler. Calls gen_event:add_handler(?MODULE, Handler, Args).

add_sup_callback(Fn)

Specs

add_sup_callback(fun((any()) -> any())) -> {ok, reference()}.

Adds a supervised callback function. The function needs to have a single argument representing the event that has been fired.

add_sup_handler(Handler, Args)

Adds a supervised event handler. Calls gen_event:add_sup_handler(?MODULE, Handler, Args).

code_change(OldVsn, State, Extra)

delete_handler(Handler)

Specs

delete_handler(module() | reference()) ->
                  term() | {error, module_not_found} | {'EXIT', Reason :: term()}.

delete_handler(Ref, Args)

Specs

delete_handler(module() | reference(), Args :: term()) ->
                  term() | {error, module_not_found} | {'EXIT', Reason :: term()}.

handle_call(Request, State)

handle_event(_, State)

handle_info(Info, State)

init(_)

notify(Event, Message)

start_link()

subscribe(EventType)

Specs

subscribe(term()) -> ok.

Subscribe to events of type Event. Any events published through update/1 will delivered to the calling process, along with all other subscribers. This function will raise an exception if you try to subscribe to the same event twice from the same process. This function uses plum_db_pubsub:subscribe/2.

subscribe(EventType, MatchSpec)

Subscribe conditionally to events of type Event. This function is similar to subscribe/2, but adds a condition in the form of an ets match specification. The condition is tested and a message is delivered only if the condition is true. Specifically, the test is: ets:match_spec_run([Msg], ets:match_spec_compile(Cond)) == [true] In other words, if the match_spec returns true for a message, that message is sent to the subscriber. For any other result from the match_spec, the message is not sent. Cond == undefined means that all messages will be delivered, which means that Cond=undefined and Cond=[{_',[],[true]}]' are equivalent. This function will raise an exception if you try to subscribe to the same event twice from the same process. This function uses plum_db_pubsub:subscribe_cond/2.

terminate(Reason, State)

unsubscribe(EventType)

Remove subscription created using subscribe/1,2

update(PObject)

Notify the event handlers, callback funs and subscribers of an updated object. The message delivered to each subscriber will be of the form: {plum_db_event, Event, Msg}