plum_db_object (plum_db v0.3.10)

Types

plum_db_context/0

Specs

plum_db_context() :: plum_db_dvvset:vector().

plum_db_modifier/0

Specs

plum_db_modifier() ::
    fun(([plum_db_value() | plum_db_tombstone()] | undefined) -> plum_db_value()).

plum_db_object/0

Specs

plum_db_object() :: {object, plum_db_dvvset:clock()}.

plum_db_tombstone/0

Specs

plum_db_tombstone() :: '$deleted'.

plum_db_value/0

Specs

plum_db_value() :: any().

Functions

context(_)

Specs

context(plum_db_object()) -> plum_db_context().

returns the context (opaque causal history) for the given object

empty_context()

Specs

empty_context() -> plum_db_context().

returns the representation for an empty context (opaque causal history)

equal_context(Context, _)

Specs

equal_context(plum_db_context(), plum_db_object()) -> boolean().

Returns true if the given context and the context of the existing object are equal

hash(_)

Specs

hash(plum_db_object()) -> binary().

returns a hash representing the objects contents

is_stale(RemoteContext, _)

Specs

is_stale(plum_db_context(), plum_db_object()) -> boolean().

Determines if the given context (version vector) is causually newer than an existing object. If the object missing or if the context does not represent an anscestor of the current key, false is returned. Otherwise, when the context does represent an ancestor of the existing object or the existing object itself, true is returned

modify(Obj, Context, Fun, ServerId)

Specs

modify(plum_db_object() | undefined,
       plum_db_context(),
       plum_db_value() | plum_db_modifier(),
       term()) ->
          plum_db_object().

modifies a potentially existing object, setting its value and updating the causual history. If a function is provided as the third argument then this function also is used for conflict resolution. The difference between this function and resolve/2 is that the logical clock is advanced in the case of this function. Additionally, the resolution functions are slightly different.

reconcile(RemoteObj, LocalObj)

Specs

reconcile(plum_db_object(), plum_db_object() | undefined) -> false | {true, plum_db_object()}.

Reconciles a remote object received during replication or anti-entropy with a local object. If the remote object is an anscestor of or is equal to the local one false is returned, otherwise the reconciled object is returned as the second element of the two-tuple

resolve(_, Reconcile)

Specs

resolve(plum_db_object(), lww | fun(([plum_db_value()]) -> plum_db_value())) -> plum_db_object().

Resolves siblings using either last-write-wins or the provided function and returns an object containing a single value. The causal history is not updated

value(Obj)

Specs

value(plum_db_object()) -> plum_db_value().

returns a single value. if the object holds more than one value an error is generated
See also: values/2.

value_count(_)

Specs

value_count(plum_db_object()) -> non_neg_integer().

returns the number of siblings in the given object

values(_)

Specs

values(plum_db_object()) -> [plum_db_value()].

returns a list of values held in the object