plum_db_partition_hashtree (plum_db v0.3.10)

Types

plum_db_key/0

Specs

plum_db_key() :: any().

plum_db_pkey/0

Specs

plum_db_pkey() :: {plum_db_prefix(), plum_db_key()}.

plum_db_prefix/0

Specs

plum_db_prefix() :: {binary() | atom(), binary() | atom()}.

Functions

code_change(OldVsn, State, Extra)

compare(Partition, RemoteFun, HandlerFun, X)

Specs

compare(plum_db:partition(), hashtree_tree:remote_fun(), hashtree_tree:handler_fun(X), X) -> X.

Compare the local tree managed by this process with the remote tree also managed by a hashtree process. RemoteFun is used to access the buckets and segments of nodes in the remote tree and should usually call get_bucket/5 and key_hashes/4. HandlerFun is used to process the differences found between the two trees. HandlerAcc is passed to the first invocation of HandlerFun. Subsequent calls are passed the return value from the previous call. This function returns the return value from the last call to HandlerFun. hashtree_tree for more details on RemoteFun, HandlerFun and HandlerAcc.

delete(PKey)

Specs

delete(plum_db_pkey()) -> ok.

Same as insert(PKey, Hash, false).

delete(Partition, PKey)

Specs

delete(pbd:partition(), plum_db_pkey()) -> ok.

Same as insert(PKey, Hash, false).

get_bucket(Node, Partition, Prefixes, Level, Bucket)

Specs

get_bucket(node(),
           non_neg_integer(),
           hashtree_tree:tree_node(),
           non_neg_integer(),
           non_neg_integer()) ->
              orddict:orddict().

Return the bucket for a node in the tree managed by this process running on Node.

handle_call(_, From, State)

handle_cast(Msg, State)

handle_info(Event, State)

init(_)

insert(PKey, Hash)

Specs

insert(plum_db_pkey(), binary()) -> ok.

Same as insert(PKey, Hash, false).

insert(PKey, Hash, IfMissing)

Specs

insert(plum_db_pkey(), binary(), boolean()) -> ok.

Insert a hash for a full-prefix and key into the tree managed by the process. If IfMissing is true the hash is only inserted into the tree if the key is not already present.

insert(Partition, PKey, Hash, IfMissing)

Specs

insert(pbd:partition(), plum_db_pkey(), binary(), boolean()) -> ok.

key_hashes(Node, Partition, Prefixes, Segment)

Specs

key_hashes(node(), non_neg_integer(), hashtree_tree:tree_node(), non_neg_integer()) ->
              orddict:orddict().

Return the key hashes for a node in the tree managed by this process running on Node.

lock(Partition)

Specs

lock(plum_db:partition()) -> ok | not_built | locked.

Locks the tree on this node for updating on behalf of the calling process.
See also: lock/3.

lock(Node, Partition)

Specs

lock(node(), plum_db:partition()) -> ok | not_built | locked.

Locks the tree on Node for updating on behalf of the calling process.
See also: lock/3.

lock(Node, Partition, Pid)

Specs

lock(node(), plum_db:partition(), pid()) -> ok | not_built | locked.

Lock the tree for updating. This function must be called before updating the tree with update/1 or update/2. If the tree is not built or already locked then the call will fail and the appropriate atom is returned. Otherwise, aqcuiring the lock succeeds and ok is returned.

name(Partition)

prefix_hash(Partition, Prefix)

Specs

prefix_hash(plum_db:partition() | pid() | atom(), plum_db_prefix() | binary() | atom()) ->
               undefined | binary().

Return the hash for the given prefix or full-prefix

release_lock(Partition)

Specs

release_lock(plum_db:partition()) -> ok.

Locks the tree on this node for updating on behalf of the calling process.
See also: lock/3.

release_lock(Node, Partition)

Specs

release_lock(node(), plum_db:partition()) -> ok.

Locks the tree on Node for updating on behalf of the calling process.
See also: lock/3.

release_lock(Node, Partition, Pid)

Specs

release_lock(node(), plum_db:partition(), pid()) -> ok | {error, not_locked | not_allowed}.

Lock the tree for updating. This function must be called before updating the tree with update/1 or update/2. If the tree is not built or already locked then the call will fail and the appropriate atom is returned. Otherwise, aqcuiring the lock succeeds and ok is returned.

reset(Partition)

Specs

reset(plum_db:partition()) -> ok.

Destroys the partition's hashtrees on the local node and rebuilds it.

reset(Node, Partition)

Specs

reset(node(), plum_db:partition()) -> ok.

Destroys the partition's hashtrees on node Node and rebuilds it.

start_link(Partition)

Specs

start_link(non_neg_integer()) -> {ok, pid()} | ignore | {error, term()}.

Starts the process using start_link/1, passing in the directory where other cluster data is stored in hashtrees_dir as the data root.

terminate(Reason, State0)

update(Partition)

Specs

update(plum_db:partition()) -> ok | not_locked | not_built | ongoing_update.

Updates the tree on this node.
See also: update/2.

update(Node, Partition)

Specs

update(node(), plum_db:partition()) -> ok | not_locked | not_built | ongoing_update.

Updates the tree on Node. The tree must be locked using one of the lock functions. If the tree is not locked or built the update will not be started and the appropriate atom is returned. Although this function should not be called without a lock, if it is and the tree is being updated by the background tick then ongoing_update is returned. If the tree is built and a lock has been acquired then the update is started and ok is returned. The update is performed asynchronously and does not block the process that manages the tree (e.g. future inserts).