plum_db_partition_server (plum_db v0.3.10)

A wrapper for an elevelb instance.

Types

db_info/0

Specs

db_info() :: #db_info{}.

iterator/0

Specs

iterator() :: #partition_iterator{}.

iterator_action/0

Specs

iterator_action() ::
    first | last | next | prev | prefetch | prefetch_stop |
    plum_db_prefix() |
    plum_db_pkey() |
    binary().

iterator_move_result/0

Specs

iterator_move_result() ::
    {ok, Key :: binary() | plum_db_pkey(), Value :: binary(), iterator()} |
    {ok, Key :: binary() | plum_db_pkey(), iterator()} |
    {error, invalid_iterator, iterator()} |
    {error, iterator_closed, iterator()} |
    {error, no_match, iterator()}.

opts/0

Specs

opts() :: [{atom(), term()}].

plum_db_key/0

Specs

plum_db_key() :: any().

plum_db_pkey/0

Specs

plum_db_pkey() :: {plum_db_prefix(), plum_db_key()}.

plum_db_prefix/0

Specs

plum_db_prefix() :: {binary() | atom(), binary() | atom()}.

plum_db_prefix_pattern/0

Specs

plum_db_prefix_pattern() ::
    {binary() | atom() | plum_db_wildcard(), binary() | atom() | plum_db_wildcard()}.

plum_db_wildcard/0

Specs

plum_db_wildcard() :: '_'.

Functions

byte_size(Id)

code_change(OldVsn, State, Extra)

erase(Partition, Key, Timeout)

get(Name, PKey)

get(Name, PKey, Opts)

get(Ref, PKey, Opts, Timeout)

handle_call(_, From, State)

handle_cast(Msg, State)

handle_info(Event, State)

init(_)

is_empty(Id)

iterator(Id, FullPrefix)

iterator(Id, FullPrefix, Opts)

If the prefix is ground, it restricts the iteration on keys belonging to that prefix and the storage type of the prefix if known. If prefix is undefined or storage type of is undefined, then it starts with disk and follows with ram. It does not cover ram_disk as all data in ram_disk is in disk but not viceversa.

iterator_close(Id, Iter)

iterator_move(Partition_iterator, Action)

Specs

iterator_move(iterator(), iterator_action()) -> iterator_move_result().

Iterates over the storage stack in order (disk -> ram_disk -> ram).

key_iterator(Id, FullPrefix)

key_iterator(Id, FullPrefix, Opts)

merge(Name, PKey, Obj)

merge(Ref, PKey, Obj, Timeout)

put(Name, PKey, ValueOrFun)

put(Name, PKey, ValueOrFun, Opts)

put(Name, PKey, ValueOrFun, Opts, Timeout)

start_link(Partition, Opts)

Specs

start_link(Partition :: non_neg_integer(), Opts :: opts()) -> any().

take(Name, PKey)

take(Name, PKey, Opts)

take(Name, PKey, Opts, Timeout)

terminate(Reason, State)