plum_db_peer_service behaviour (plum_db v0.3.10)

Based on: github.com/lasp-lang/lasp/...lasp_peer_service.erl

Types

partisan_peer/0

Specs

partisan_peer() :: #{name := _Name, listen_addrs := _ListenAddrs}.

Callbacks

connections/0

Specs

connections() -> {ok, [partisan_peer_service_connections:t()]}.

decode/0

Specs

decode() -> list().

join/1

Specs

join(node() | list() | partisan_peer()) -> ok | {error, atom()}.

join/2

Specs

join(node() | list() | partisan_peer(), boolean()) -> ok | {error, atom()}.

join/3

Specs

join(node() | list() | partisan_peer(), node() | list() | partisan_peer(), boolean()) ->
        ok | {error, atom()}.

leave/0

Specs

leave() -> ok.

leave/1

Specs

leave(node() | list() | partisan_peer()) -> ok.

manager/0

Specs

manager() -> module().

members/0

Specs

members() -> {ok, [node()]}.

mynode/0

Specs

mynode() -> atom().

myself/0

Specs

myself() -> map().

stop/0

Specs

stop() -> ok.

stop/1

Specs

stop(iolist()) -> ok.

sync_join/1

Specs

sync_join(node() | list() | partisan_peer()) -> ok | {error, atom()}.

sync_join/2

Specs

sync_join(node() | list() | partisan_peer(), boolean()) -> ok | {error, atom()}.

sync_join/3

Specs

sync_join(node() | list() | partisan_peer(), node() | list() | partisan_peer(), boolean()) ->
             ok | {error, atom()}.

Functions

add_sup_callback(Function)

Add callback

cast_message(Name, ServerRef, Message)

Add callback

connections()

Return node connections.

decode(State)

forward_message(Name, ServerRef, Message)

Add callback

join(Nodename)

Prepare node to join a cluster.

join(NodeStr, Auto)

Convert nodename to atom.

join(Node1, Node2, Auto)

Initiate join. Nodes cannot join themselves.

leave()

Leave the cluster.

leave(Peer)

Leave the cluster.

manager()

Return manager.

members()

Return cluster members.

mynode()

Return myself.

myself()

Return myself.

peer(Nodename)

Returns a local peer node i.e. running on the same host, with nodename Name and same metadata as myself/0. You should only use this function for testing purposes as there is no much sense in running a plum_db cluster on the same host.

peer(Nodename, Term)

Returns a peer with nodename Name and host 'Host' and same metadata as myself/0.

peer_service()

Return the currently active peer service.

stop()

Stop node.

stop(Reason)

Stop node for a given reason.

sync_join(Node)

Prepare node to join a cluster.

sync_join(Node, Auto)

Prepare node to join a cluster.

sync_join(Node1, Node2, Auto)

Prepare node to join a cluster.

update_members(Nodes)

Update cluster members.